Create Final_Project.py
Raspberry pi  preloaded with Python ,a Python integrated development environment is the main programming language for raspberry pi . There are bindings in other languages. Raspberry pi runs on a variety of platforms. The desktop which can be used is Windows, Linux, FreeBSD, OpenBSD, Real time OS ansd in the mobile such as Andriod, ios, Maemo.